re-home Jinja plugin decorators public API
