+++ lib/ansible/_internal/ansible_collections/ansible/_protomatter/plugins/filter/true_type.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import typing as t
 
-from ansible.plugins import accept_args_markers
+from ansible.template import accept_args_markers
 
 
 @accept_args_markers
